Semantics: Environment: added environment definition.
Environment (also, valuation) is a function from the set of all
variables to the domain of a first-order structure.

Defined extensions of environment to terms and vectors. Tried to prove
a lemma from Brotherstons PhD thesis, but got stuck in the inductive
case. Do I need an even better induction principle, or should I learn
about induction and destruction of dependent types better?