Syntax: Term: added another induction principle, proved lemmas.
Also added a helper lemma for working with vectors. Possible refactor:
move such helper lemmas and module declarations into a base
file. Since we are working with [Vector]s and [Ensemble]s, each of
which have their own [In] and [Forall] versions, it would be a good
thing to prefix all occurrences of such functions with [V.] and [E.],
mainly for readability and consistency.