Theory: started work on the final rule.
Defined finite substitution (identity on all but a finite number of
elements, defined by input vectors). Proved some lemmas about it.
Formalized part of the left rule. TODO: prettify, it is unreadable at
the moment. TODO: for each minor premise, determine max var which
appears in any of the premises and shift gamma and delta by it.